Implement load and save in the core
Implement load and save functionality in the SDAM core. This implementation is quite memory inefficient, requiring three copies of the audio data to exist in the RAM at one point of the process, and needs testing to determine whether it works at all.
Detailed audio data space requirement measurements need to be taken in order to judge the relevance of these flaws.