Implement loading and saving in GUI and TUI
Implement loading and saving in the terminal interface. Since this interface is intended for testing purposes only, there are just commands for saving into / loading from a file with hardcoded name.
Implement a proper user-friendly loading / saving in the GUI. There are menu items, a save keyboard shortcut (dialog is opened only if the program is working with a new memory), loading can be invoked through a CLI argument and shows an error dialog if something goes wrong.
Exception handling between Rust and Python is something that will require more work, for now empty / filled strings are used, but PyO3 has apparently better ways of solving this.
Fix few issues in the core.