Rewrite the structure of PlaybackState
Previously, PlaybackState enum had carried the instances of a cpal stream and audio producer responsible for sending output audio data. The reason was that these were initialized if and only if the state was set to Playing or Paused, not Stopped, which was supposed to represent a program state where no output audio playback is active, playing or paused. The reason for attaching these instances to the enum was to avoid having to check them via an if let statement for initialization, when the code already needed to check whether the playback state is playing or paused.
However, this architecture also brought challenges:
* When swapping an instance of the enum itself for a different variant-one, it was necessary to make sure the attached objects won't get dropped. This required them to be wrapped in Arcs, that needed to be cloned before doing a swap.
* Since Arc content is read-only and audio-producer is an object that only makes sense with a mutable acces, it required a mutex, which needed to be also recloned in the update audio message handler of AudioHandler for borrowing issues.

Since PlaybackState::Stopped no longer exists, this solution became unnecessary and got removed in this commit.
The output stream and audio producer now live as fields of AudioHandler, while PlaybackState is (for now) without any parameters.