Implement time travel
Implement time travel functionality in the GUI. It can be activated from non-recording as well as recording state, which is persisted after the mode is deactivated, unless stop recording is used specifically, which will cancel both timetravel and the recording itself.
One restriction had to be placed on playback, which should be now unavailable if recording is active, and can be used only through the time travel mode. The reason for this is the mark placing proces. Normally, when recording is active and playback is paused, the cursor is not moving and marks are automatically placed on the end of the available audio, because that's where the user likely wants to place them, as this feature is intended to be used for marking semantic on the go.
However, when the user pauses the playback in timetravel mode, the approach is different, now the target for new marks is the current audio cursor position, even though recording is still ongoing in the background.
if the user was able to freely use playback during recording, it would be impossible to distinguish these two use-cases, therefore it is possible only through the timetravel functionality. Which is equally capable to using the standard playback, so there shouldn't be any functional restrictions.