Incorporate the Python GUi into a Briefcase template for packaging
Restructure the gui_py code. Now, the root gui_py directory contains directories for several packages - gui_py (the Rust interface to SDAM), sdam (the actual briefcase application), speechd (an interface to Speech Dispatcher) and the .gitignore also leaves place for venv and wheels folders to be placed here.
Add speechd package, this is a poetry library incorporating the Brailcom API to Speech dispatcher. It's a very proof of concept packaging, which currently lacks tests and should eventually go into its own repo and be properly submitted to PyPI, for now, it resides here to support speech in the SDAM project.